refactor(resolveBody): 按解析路径拆分为 resolveBody/ 子模块
- shared: ResolvedBody 类型与 HTML 清洗/转义/图片绝对化工具
- guards: 反爬/付费墙识别、摘要 Feed 判定、软降级正文
- extractors: Readability 主路径 + 优设/公众号站点定制抽取
- siteBodies: 虎嗅/网易/知乎/机器之心详情接口正文
- video: 视频占位正文与媒体嗅探增量更新
- lib/resolveBody.ts 保留为编排入口(resolveArticleBody), 外部导入路径不变
